refactor(rviz): Panel の config/ComboBox 系を別 TU へ分割 (#397)
